+++ app/domain/authentication/authn_jwt/signing_key/fetch_cached_signing_key.rb
@@ -6,20 +6,18 @@ module Authentication
       # fetch_signing_key it is extreme case that error need to be raised so it can be investigated so reek will ignore
       # this.
       # :reek:InstanceVariableAssumption
-      class FetchCachedSigningKey
-        def initialize(logger: Rails.logger)
-          @logger = logger
-        end
-
-        def call(signing_key_interface:)
-          @signing_key_interface = signing_key_interface
+      FetchCachedSigningKey = CommandClass.new(
+        dependencies: {},
+        inputs: %i[signing_key_provider]
+      ) do
+        def call
           fetch_signing_key
         end
 
         private
 
         def fetch_signing_key
-          @signing_key_interface.fetch_signing_key
+          @signing_key_provider.fetch_signing_key
         end
       end
     end