+++ app/domain/authentication/authn_jwt/validate_status.rb
@@ -3,7 +3,7 @@ module Authentication
 
     ValidateStatus = CommandClass.new(
       dependencies: {
-        fetch_signing_key_from_cache: ::Util::ConcurrencyLimitedCache.new(
+        fetch_signing_key: ::Util::ConcurrencyLimitedCache.new(
           ::Util::RateLimitedCache.new(
             ::Authentication::AuthnJwt::SigningKey::FetchCachedSigningKey.new,
             refreshes_per_interval: CACHE_REFRESHES_PER_INTERVAL,
@@ -13,7 +13,7 @@ module Authentication
           max_concurrent_requests: CACHE_MAX_CONCURRENT_REQUESTS,
           logger: Rails.logger
         ),
-        create_signing_key_interface: Authentication::AuthnJwt::SigningKey::CreateSigningKeyFactory.new,
+        create_signing_key_provider: Authentication::AuthnJwt::SigningKey::CreateSigningKeyProvider.new,
         fetch_issuer_value: Authentication::AuthnJwt::ValidateAndDecode::FetchIssuerValue.new,
         fetch_audience_value: Authentication::AuthnJwt::ValidateAndDecode::FetchAudienceValue.new,
         fetch_enforced_claims: Authentication::AuthnJwt::RestrictionValidation::FetchEnforcedClaims.new,
@@ -144,15 +144,15 @@ module Authentication
       end
 
       def validate_signing_key
-        @fetch_signing_key_from_cache.call(
-          cache_key: signing_key_interface.signing_key_uri,
-          signing_key_interface: signing_key_interface
+        @fetch_signing_key.call(
+          cache_key: signing_key_provider.signing_key_uri,
+          signing_key_provider: signing_key_provider
         )
         @logger.debug(LogMessages::Authentication::AuthnJwt::ValidatedSigningKeyConfiguration.new)
       end
       
-      def signing_key_interface
-        @signing_key_interface ||= @create_signing_key_interface.call(
+      def signing_key_provider
+        @signing_key_provider ||= @create_signing_key_provider.call(
           authentication_parameters: authentication_parameters
         )
       end