+++ spec/app/domain/authentication/authn-jwt/validate_and_decode/validate_and_decode_token_spec.rb
@@ -45,16 +45,16 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
     )
   }
 
-  let(:mocked_signing_key_interface_factory_valid) { double("MockedSigningKeyInterfaceFactoryValid") }
-  let(:mocked_signing_key_interface_factory_invalid) { double("MockedSigningKeyInterfaceFactoryInvalid") }
-  let(:mocked_signing_key_interface_factory_failed) { double("MockedSigningKeyInterfaceFactoryFailed") }
+  let(:mocked_create_signing_key_provider_valid) { double("MockedSigningKeyInterfaceFactoryValid") }
+  let(:mocked_create_signing_key_provider_invalid) { double("MockedSigningKeyInterfaceFactoryInvalid") }
+  let(:mocked_create_signing_key_provider_failed) { double("MockedSigningKeyInterfaceFactoryFailed") }
 
-  let(:signing_key_interface_factory_error) { "signing key interface factory error" }
+  let(:create_signing_key_provider_error) { "signing key interface factory error" }
 
-  let(:mocked_fetch_signing_key_interface_valid) { double("MockedSigningKeyInterfaceValid") }
-  let(:mocked_fetch_signing_key_interface_failed) { double("MockedSigningKeyInterfaceFailed") }
+  let(:mocked_fetch_signing_key_provider_valid) { double("MockedSigningKeyInterfaceValid") }
+  let(:mocked_fetch_signing_key_provider_failed) { double("MockedSigningKeyInterfaceFailed") }
 
-  let(:fetch_signing_key_interface_error) { "fetch signing key interface error" }
+  let(:fetch_signing_key_provider_error) { "fetch signing key interface error" }
 
   let(:mocked_fetch_signing_key_failed_on_1st_time) { double("MockedFetchSigningKeyInvalid") }
   let(:mocked_fetch_signing_key_failed_on_2nd_time) { double("MockedFetchSigningKeyInvalid") }
@@ -136,31 +136,31 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
   let(:mocked_verify_and_decode_token_succeed_to_validate_claims_when_keys_updated) { double("MockedVerifyAndDecodeTokenSucceedToValidateClaims") }
 
   before(:each) do
-    allow(mocked_fetch_signing_key_interface_valid).to(
+    allow(mocked_fetch_signing_key_provider_valid).to(
       receive(:signing_key_uri).and_return(valid_signing_key_uri)
     )
 
-    allow(mocked_signing_key_interface_factory_valid).to(
-      receive(:call).and_return(mocked_fetch_signing_key_interface_valid)
+    allow(mocked_create_signing_key_provider_valid).to(
+      receive(:call).and_return(mocked_fetch_signing_key_provider_valid)
     )
 
-    allow(mocked_fetch_signing_key_interface_failed).to(
-      receive(:signing_key_uri).and_raise(fetch_signing_key_interface_error)
+    allow(mocked_fetch_signing_key_provider_failed).to(
+      receive(:signing_key_uri).and_raise(fetch_signing_key_provider_error)
     )
 
-    allow(mocked_signing_key_interface_factory_invalid).to(
-      receive(:call).and_return(mocked_fetch_signing_key_interface_failed)
+    allow(mocked_create_signing_key_provider_invalid).to(
+      receive(:call).and_return(mocked_fetch_signing_key_provider_failed)
     )
 
-    allow(mocked_signing_key_interface_factory_failed).to(
-      receive(:call).and_raise(signing_key_interface_factory_error)
+    allow(mocked_create_signing_key_provider_failed).to(
+      receive(:call).and_raise(create_signing_key_provider_error)
     )
 
     allow(mocked_fetch_signing_key_failed_on_1st_time).to(
       receive(:call).with(
         refresh: false,
         cache_key: anything(),
-        signing_key_interface: anything()
+        signing_key_provider: anything()
       ).and_raise(fetch_signing_key_1st_time_error)
     )
 
@@ -168,7 +168,7 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       receive(:call).with(
         refresh: false,
         cache_key: anything(),
-        signing_key_interface: anything()
+        signing_key_provider: anything()
       ).and_return(jwks_from_2nd_call)
     )
 
@@ -176,7 +176,7 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       receive(:call).with(
         refresh: true,
         cache_key: anything(),
-        signing_key_interface: anything()
+        signing_key_provider: anything()
       ).and_raise(fetch_signing_key_2nd_time_error)
     )
 
@@ -188,7 +188,7 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       receive(:call).with(
         refresh: false,
         cache_key: anything(),
-        signing_key_interface: anything()
+        signing_key_provider: anything()
       ).and_return(jwks_from_1st_call)
     )
 
@@ -196,7 +196,7 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       receive(:call).with(
         refresh: true,
         cache_key: anything(),
-        signing_key_interface: anything()
+        signing_key_provider: anything()
       ).and_return(jwks_from_2nd_call)
     )
 
@@ -345,38 +345,38 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
     context "When error is during signing key factory call" do
       subject do
         ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-          fetch_signing_key_from_cache: mocked_fetch_signing_key_failed_on_1st_time,
-          signing_key_interface_factory: mocked_signing_key_interface_factory_failed
+          fetch_signing_key: mocked_fetch_signing_key_failed_on_1st_time,
+          create_signing_key_provider: mocked_create_signing_key_provider_failed
         ).call(
           authentication_parameters: authentication_parameters_with_valid_token
         )
       end
 
       it "raises an error" do
-        expect { subject }.to raise_error(signing_key_interface_factory_error)
+        expect { subject }.to raise_error(create_signing_key_provider_error)
       end
     end
 
     context "When error is during signing_key_uri call" do
       subject do
         ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-          fetch_signing_key_from_cache: mocked_fetch_signing_key_failed_on_1st_time,
-          signing_key_interface_factory: mocked_signing_key_interface_factory_invalid
+          fetch_signing_key: mocked_fetch_signing_key_failed_on_1st_time,
+          create_signing_key_provider: mocked_create_signing_key_provider_invalid
         ).call(
           authentication_parameters: authentication_parameters_with_valid_token
         )
       end
 
       it "raises an error" do
-        expect { subject }.to raise_error(fetch_signing_key_interface_error)
+        expect { subject }.to raise_error(fetch_signing_key_provider_error)
       end
     end
 
     context "When error is during fetching from cache" do
       subject do
         ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-          fetch_signing_key_from_cache: mocked_fetch_signing_key_failed_on_1st_time,
-          signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+          fetch_signing_key: mocked_fetch_signing_key_failed_on_1st_time,
+          create_signing_key_provider: mocked_create_signing_key_provider_valid
         ).call(
           authentication_parameters: authentication_parameters_with_valid_token
         )
@@ -393,9 +393,9 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       context "and failed to fetch keys from provider" do
         subject do
           ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-            fetch_signing_key_from_cache: mocked_fetch_signing_key_failed_on_2nd_time,
+            fetch_signing_key: mocked_fetch_signing_key_failed_on_2nd_time,
             verify_and_decode_token: mocked_verify_and_decode_token_invalid,
-            signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+            create_signing_key_provider: mocked_create_signing_key_provider_valid
           ).call(
             authentication_parameters: authentication_parameters_with_valid_token
           )
@@ -409,9 +409,9 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       context "and succeed to fetch keys from provider" do
         subject do
           ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-            fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+            fetch_signing_key: mocked_fetch_signing_key_always_succeed,
             verify_and_decode_token: mocked_verify_and_decode_token_invalid,
-            signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+            create_signing_key_provider: mocked_create_signing_key_provider_valid
           ).call(
             authentication_parameters: authentication_parameters_with_valid_token
           )
@@ -427,11 +427,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       context "and keys are not updated" do
         subject do
           ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-            fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+            fetch_signing_key: mocked_fetch_signing_key_always_succeed,
             verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_2nd_time,
             fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
             get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_valid,
-            signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+            create_signing_key_provider: mocked_create_signing_key_provider_valid
           ).call(
             authentication_parameters: authentication_parameters_with_valid_token
           )
@@ -445,11 +445,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       context "and keys are updated" do
         subject do
           ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-            fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+            fetch_signing_key: mocked_fetch_signing_key_always_succeed,
             verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_1st_time,
             fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
             get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_valid,
-            signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+            create_signing_key_provider: mocked_create_signing_key_provider_valid
           ).call(
             authentication_parameters: authentication_parameters_with_valid_token
           )
@@ -467,10 +467,10 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
       context "and failed to fetch enforced claims" do
         subject do
           ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-            fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+            fetch_signing_key: mocked_fetch_signing_key_always_succeed,
             verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_1st_time,
             fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_invalid,
-            signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+            create_signing_key_provider: mocked_create_signing_key_provider_valid
           ).call(
             authentication_parameters: authentication_parameters_with_valid_token
           )
@@ -485,10 +485,10 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
         context "with empty claims list to validate" do
           subject do
             ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-              fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+              fetch_signing_key: mocked_fetch_signing_key_always_succeed,
               verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_1st_time,
               fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_with_empty_claims,
-              signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+              create_signing_key_provider: mocked_create_signing_key_provider_valid
             ).call(
               authentication_parameters: authentication_parameters_with_valid_token
             )
@@ -502,10 +502,10 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
         context "with mandatory claims which do not exist in token" do
           subject do
             ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-              fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+              fetch_signing_key: mocked_fetch_signing_key_always_succeed,
               verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_1st_time,
               fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_with_not_exist_claims_in_token,
-              signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+              create_signing_key_provider: mocked_create_signing_key_provider_valid
             ).call(
               authentication_parameters: authentication_parameters_with_valid_token
             )
@@ -519,11 +519,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
         context "and failed to get verification options" do
           subject do
             ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-              fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+              fetch_signing_key: mocked_fetch_signing_key_always_succeed,
               verify_and_decode_token: mocked_verify_and_decode_token_succeed_on_1st_time,
               fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
               get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_invalid,
-              signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+              create_signing_key_provider: mocked_create_signing_key_provider_valid
             ).call(
               authentication_parameters: authentication_parameters_with_valid_token
             )
@@ -544,11 +544,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
           context "and failed to validate claims" do
             subject do
               ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-                fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+                fetch_signing_key: mocked_fetch_signing_key_always_succeed,
                 verify_and_decode_token: mocked_verify_and_decode_token_failed_to_validate_claims,
                 fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
                 get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_valid,
-                signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+                create_signing_key_provider: mocked_create_signing_key_provider_valid
               ).call(
                 authentication_parameters: authentication_parameters_with_valid_token
               )
@@ -563,11 +563,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
             context "and keys are not updated" do
               subject do
                 ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-                  fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+                  fetch_signing_key: mocked_fetch_signing_key_always_succeed,
                   verify_and_decode_token: mocked_verify_and_decode_token_succeed_to_validate_claims_when_keys_not_updated,
                   fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
                   get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_valid,
-                  signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+                  create_signing_key_provider: mocked_create_signing_key_provider_valid
                 ).call(
                   authentication_parameters: authentication_parameters_with_valid_token
                 )
@@ -581,11 +581,11 @@ RSpec.describe('Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeTo
             context "and keys are updated" do
               subject do
                 ::Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken.new(
-                  fetch_signing_key_from_cache: mocked_fetch_signing_key_always_succeed,
+                  fetch_signing_key: mocked_fetch_signing_key_always_succeed,
                   verify_and_decode_token: mocked_verify_and_decode_token_succeed_to_validate_claims_when_keys_updated,
                   fetch_jwt_claims_to_validate: mocked_fetch_jwt_claims_to_validate_valid,
                   get_verification_option_by_jwt_claim: mocked_get_verification_option_by_jwt_claim_valid,
-                  signing_key_interface_factory: mocked_signing_key_interface_factory_valid
+                  create_signing_key_provider: mocked_create_signing_key_provider_valid
                 ).call(
                   authentication_parameters: authentication_parameters_with_valid_token
                 )