+++ spec/app/domain/authentication/authn-jwt/validate_status_spec.rb
@@ -23,8 +23,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
   }
 
   let(:mocked_logger) { double("Mocked logger")  }
-  let(:mocked_valid_create_signing_key_interface) { double("Mocked valid create signing key interface")  }
-  let(:mocked_invalid_create_signing_key_interface) { double("Mocked invalid create signing key interface")  }
+  let(:mocked_valid_create_signing_key_provider) { double("Mocked valid create signing key interface")  }
+  let(:mocked_invalid_create_signing_key_provider) { double("Mocked invalid create signing key interface")  }
   let(:mocked_valid_fetch_issuer_value) { double("Mocked valid fetch issuer value")  }
   let(:mocked_invalid_fetch_issuer_value) { double("Mocked invalid fetch issuer value")  }
   let(:mocked_invalid_fetch_audience_value) { double("Mocked invalid audience issuer value")  }
@@ -54,24 +54,24 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
   let(:user_cant_access_webservice_error) { "User cant access webservice" }
   let(:webservice_does_not_exist_error) { "Webservice does not exist" }
   let(:account_does_not_exist_error) { "Account does not exist" }
-  let(:mocked_valid_signing_key_interface) { double("Mocked valid signing key interface")  }
-  let(:mocked_valid_fetch_signing_key_from_cache) { double("Mocked valid fetch signing key interface")  }
+  let(:mocked_valid_signing_key_provider) { double("Mocked valid signing key interface")  }
+  let(:mocked_valid_fetch_signing_key) { double("Mocked valid fetch signing key interface")  }
 
 
   before(:each) do
-    allow(mocked_valid_create_signing_key_interface).to(
-      receive(:call).and_return(mocked_valid_signing_key_interface)
+    allow(mocked_valid_create_signing_key_provider).to(
+      receive(:call).and_return(mocked_valid_signing_key_provider)
     )
 
-    allow(mocked_valid_signing_key_interface).to(
+    allow(mocked_valid_signing_key_provider).to(
       receive(:signing_key_uri).and_return(valid_signing_key_uri)
     )
 
-    allow(mocked_valid_fetch_signing_key_from_cache).to(
+    allow(mocked_valid_fetch_signing_key).to(
       receive(:call).and_return(valid_signing_key)
     )
     
-    allow(mocked_invalid_create_signing_key_interface).to(
+    allow(mocked_invalid_create_signing_key_provider).to(
       receive(:call).and_raise(create_signing_key_configuration_is_invalid_error)
     )
 
@@ -166,8 +166,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
       subject do
         ::Authentication::AuthnJwt::ValidateStatus.new(
-          fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-          create_signing_key_interface: mocked_valid_create_signing_key_interface,
+          fetch_signing_key: mocked_valid_fetch_signing_key,
+          create_signing_key_provider: mocked_valid_create_signing_key_provider,
           fetch_issuer_value: mocked_valid_fetch_issuer_value,
           identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
           validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -191,8 +191,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -215,8 +215,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -239,8 +239,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -263,8 +263,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_invalid_validate_webservice_is_whitelisted,
@@ -299,8 +299,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
 
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -324,8 +324,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "signing key secrets are not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_invalid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_invalid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -347,8 +347,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "issuer secrets are not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_invalid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,
@@ -370,8 +370,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "audience secret is not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             fetch_audience_value: mocked_invalid_fetch_audience_value,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
@@ -394,8 +394,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "enforced claims is not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             fetch_enforced_claims: mocked_invalid_fetch_enforced_claims,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
@@ -418,8 +418,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "mapping claims is not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             fetch_mapping_claims: mocked_invalid_fetch_mapping_claims,
             identity_from_decoded_token_provider_class: mocked_valid_identity_from_decoded_token_provider,
@@ -442,8 +442,8 @@ RSpec.describe('Authentication::AuthnJwt::ValidateStatus') do
       context "identity secrets are not configured properly" do
         subject do
           ::Authentication::AuthnJwt::ValidateStatus.new(
-            fetch_signing_key_from_cache: mocked_valid_fetch_signing_key_from_cache,
-            create_signing_key_interface: mocked_valid_create_signing_key_interface,
+            fetch_signing_key: mocked_valid_fetch_signing_key,
+            create_signing_key_provider: mocked_valid_create_signing_key_provider,
             fetch_issuer_value: mocked_valid_fetch_issuer_value,
             identity_from_decoded_token_provider_class: mocked_invalid_identity_from_decoded_token_provider,
             validate_webservice_is_whitelisted: mocked_valid_validate_webservice_is_whitelisted,